backend/requirements.txt
Flask
flask-cors
crewai
crewai-tools
uv
openai-whisper
setuptools-rust